Refactor: delete tsx에서 svg로 변경
