+++ src/app/custom-keyboard/_components/ColorTag.tsx
@@ -3,7 +3,7 @@
 import classNames from 'classnames/bind';
 import type { CustomKeyboardKeyTypes } from '@/types/CustomKeyboardTypes';
 import { Color } from '@react-three/fiber';
-import DeleteTagIcon from '@/public/svgs/delete';
+import DeleteTagIcon from '@/public/svgs/delete.svg';
 import styles from './ColorTag.module.scss';
 
 const cn = classNames.bind(styles);
@@ -18,7 +18,7 @@ export default function ColorTag({ keyCap, color, onDelete }: ColorTagProps) {
   return (
     <div className={cn('wrapper')}>
       <div>{`${keyCap}: ${color}`}</div>
-      <DeleteTagIcon width={20} height={20} onClick={onDelete} />
+      <DeleteTagIcon width={20} height={20} fill='#4968F6' onClick={onDelete} />
     </div>
   );
 }
